django_backend/requirements.txt
# Django Backend Requirements
# Production dependencies

Django==5.0.2
djangorestframework==3.14.0
psycopg2-binary==2.9.9
django-cors-headers==4.3.1
python-dotenv==1.0.1

# Optional for production
gunicorn==21.2.0
whitenoise==6.6.0
